+++ de.lproj/Localizable.strings
@@ -102,9 +102,17 @@
 "interval.fifteen.minutes"="Fünfzehn Minutes";
 "interval.thirty.minutes"="Dreißig Minutes";
 "interval.one.hour"="Eine Stunde";
+"interval.two.hours"="Two Hours";
+"interval.three.hours"="Three Hours";
+"interval.four.hours"="Four Hours";
+"interval.five.hours"="Five Hours";
 "interval.six.hours"="Sechs Stunden";
 "interval.twelve.hours"="Zwölf Stunden";
+"interval.eighteen.hours"="Eighteen Hours";
 "interval.twentyfour.hours"="Vierundzwanzig Stunden";
+"interval.thirtysix.hours"="Thirty Six Hours";
+"interval.fortyeight.hours"="Forty Eight Hours Hours";
+"interval.seventytwo.hours"="Seventy Two Hours";
 "keyboard.type"="Keyboard Typ";
 "logging"="Logging";
 "lora"="LoRa";
